requirements.txt
numpy
pyzed.sl
cv2
glob